+++ src/AmpControl/StereoVolume.cpp
@@ -6,10 +6,10 @@ StereoVolume::~StereoVolume() { }
 String StereoVolume::getName() { return name; }
 void StereoVolume::setName(String name) { this->name = name; }
 
-float StereoVolume::getOffsetLeft() { return offsetLeft; }
-float StereoVolume::getOffsetRight() { return offsetRight; }
+int StereoVolume::getOffsetLeft() { return offsetLeft; }
+int StereoVolume::getOffsetRight() { return offsetRight; }
 
-void StereoVolume::setOffset(float offsetLeft, float offsetRight)
+void StereoVolume::setOffset(int offsetLeft, int offsetRight)
 {
     this->offsetLeft = offsetLeft;
     this->offsetRight = offsetRight;
